ART-Extension: Trivial Loop Evaluator
This patch allows to statically compute the result of trivial loops. It is
especially useful for breaking the Quadrant CPU benchmark. A loop can be
evaluated by TLE if all of its inputs are constant and if the loop is easy
enough to deal with.

Change-Id: I6b7fc3bacea7dcef0dd14992be5b088bae43dd46
Category: Device Enablement
Domain: AOSP.ART-ME
Origin: Internal
Upstream-Candidate: no, proprietary
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-1142
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-129
Signed-off-by: Jean-Philippe Halimi <jean-philippe.halimi@intel.com>
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>