Generate Select instructions for X86 code gen
Add a pass that changes small if diamonds (0 or 1 instruction in each
side of the diamond) feeding a Phi, replacing the Phi with an
HX86Select instruction that uses CMOV to avoid branching code.

- The comparision on the HIf must not be float or double.
- In 32 bit mode, the comparison must not be long.
- The values being assigned must not be float or double.
- Any instructions in the true/false blocks must not be volatile.

This handles simple triangles and diamonds in the CFG.  The current
limitation is that there can be only 1 Phi in the merge block.

Squash:
Icb6d978 Fix the GenerateSelect to reject throwable instructions

Change-Id: I69e3e9f54070a3289e4f918b9a0500d70a8119aa
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-1382
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-3999
Category: device enablement
Domain: AOSP.ART-BE
Origin: internal
Upstream-Candidate: no, not-aosp
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>