Fix Non Temporal Move with loads in loop
The document for generating non temporal moves stated that there must be
no loads in the loop.  The code in the optimizer that tried to detect
that was incorrect.

There is actually no need to worry about instance or static loads in the
loop, as the memory cannot overlap with array data.

Accordingly, I have updated the QA document to state that only array
loads in the loop prevent this optimization, rather than all loops, and
I have updated the source to detect them properly.  The generated
message in this case is also updated.

Change-Id: I25bad6e3daf100152bdbd2075fa521462f7cbb55
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-2821
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-9801
Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Upstream-Candidate: no, not-aosp
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>