ART: Retry log message delivery in case of errors
If too many messages are logged during a short period of time
some of them might be lost. This can result in failures for
applications that read and analyze the log messages.

This patch introduces a process specific environment variable

  DALVIK_LOG_RETRY_DELIVERY=<retry_count>/<delay_millis>

If it is set then an ART process (dalvik, dex2oat, patchoat, ...)
tries sending a log message the specified number of times with the
specified delay doubled every next try.

On the aosp_x86_64 emulator even the minimal settings
  DALVIK_LOG_RETRY_DELIVERY=1/0
are enough if only one process is actively logging.

This patch is an ART local solution of the message loss problem
described in.
https://android-review.googlesource.com/131020 liblog: prevent message loss by limiting throughput
https://android-review.googlesource.com/148432 liblog: retry log message delivery in case of errors
https://code.google.com/p/android/issues/detail?id=173269 fast printed Android log messages are lost

The main goal of this patch to persuade Google to merge it in into
AOSP. Otherwise we will not be able to merge it into all our
workspaces and in this case the patch is useless for QA needs.

Change-Id: Iba255fa429843d1b64bd6f14fbb4d9208942a418
Category: device enablement
Domain: AOSP.ART-Other
Origin: internal
Upstream-Candidate: yes
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/OAM-1850
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-9801
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>
Signed-off-by: Serdjuk, Nikolay Y <nikolay.y.serdjuk@intel.com>