ART-Extension: Fix LHSS with Phi of different types.
LHSS creates a Phi node with two inputs: getter and value to set.
If value to set is of fp type and prtimitive type propagation
was not able to prove that, then Phi node will have inputs of
different types. This patch fixes this by changing the constant
of value to set to the corresponding fp constant.

This is bug fix with no performance impact expected.

Change-Id: I9872594ec1e5e28faa05301c544758cc46c4a4b6
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-2838
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-9801
Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>
Signed-off-by: Serdjuk, Nikolay Y <nikolay.y.serdjuk@intel.com>