ART-Extension: Fix VPTH optimization
VPTH does not expect that there are several setters to
the same location in pre-header. This causes the crash of the
compiler.

This patch takes only the last setter to the same location and
uses it for optimization.

No performance impact is expected.

Change-Id: I3f6f1a329c3bdd6b833b481507475fba9f97db76
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-2890
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-9801
Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>
Signed-off-by: Serdjuk, Nikolay Y <nikolay.y.serdjuk@intel.com>