ART-Extension: Enable peeling for LoadClass and LoadString
LoadString and LoadClass have no IR inputs, are not marked as throwers,
and also they are not marked as being heap readers (since they
load things not related to current method). However, these
characteristics made it so that they were not being considered in
peeling. However, they are useful to peel because the values they load
are invariants and it does not make sense to do the loading in loop.

This example is now supported for peeling (and was not before):
  public static void loop() {
    int sum = 0;
    for (int i = 0; i < 500; i++){
      sum += Cls.field1;
    }
  }

Expected-Impact: Increases the application of peeling marginally by
handling a few additional cases which are "heap readers" but were not
categorized as so. No performance impact is expected in general case.

Change-Id: I6ce63a799545f10d07f6a98f5166da21e3be9ff1
Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Upstream-Candidate: no, proprietary
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-2982
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-15426
Signed-off-by: Razvan A Lupusoru <razvan.a.lupusoru@intel.com>
Signed-off-by: Serdjuk, Nikolay Y <nikolay.y.serdjuk@intel.com>