ART-Extension: Improve cloner for better peeling, FBL, unrolling
The cloner has been updated to support better dex pc propagation. It can
also clone the following instructions which were not supported before:
- NewArray, NewInstance
- InvokeStaticOrDirect, InvokeVirtual, InvokeInterface
- MonitorOperation, MemoryBarrier
- LoadString, Throw, Select, Return
- Suspend, TestSuspend
- Unresolved invokes, field getters and setters
In addition to supporting those nodes above, all operations that were
previously supported and defined an object, now receive a copy of the
type propagation.

It also adds more sanity checks like instruction data equality and
environment checking. It also makes it so that in debug mode, unclonable
node is a fatal failure.

In order to limit peeling a bit now that cloner is no longer an issue,
the peeling has been enhanced to only peel if instruction can be moved.
This makes it so instructions like invokes (which cannot be yet proved
if two calls are same with no side-effects) can now be ignored when
considering peeling. This is similar to old behavior (which would reject
due to cloning failure). It also relaxes the BB count requirement in
peeling in favor of instruction counts in order to reduce impact from
code bloating.

Additionally, loop unrolling and form bottom loop (FBL) passes were
updated to reject try/catch blocks and irreducible loops since cloning
made it possible to accept them.

Expected-Impact: There is no performance impact expected here but cloner
is generalized to support more cases. This makes it so that passes that
rely on it do not reject because cloning is not supported. This means
that peeling, unrolling, and bottom loop formation should be more
applicable. This is shown in the following cases.
- Caffeinemark
 Peeling: 3 -> 5
 Unrolling: 1 -> 1 (no change)
 FBL: 8 -> 18
- IcyRocks
 Peeling: 40 -> 164
 Unrolling: 0 -> 8
 FBL: 405 -> 505
- core-optimizing.oat
 Peeling: 556 -> 1040
 Unrolling: 13 -> 25
 FBL: 2381 -> 3117

Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Upstream-Candidate: no, proprietary
Tracked-On: https://jira01.devtools.intel.com/browse/CAR-3063
Change-Id: I23faceeb1e220e026d89ab58a9fe673fe513198d
Signed-off-by: Razvan A Lupusoru <razvan.a.lupusoru@intel.com>
Reviewed-on: https://android.intel.com/478527
Reviewed-by: Popov, Ivan G <ivan.g.popov@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>