ART-Extension: HLoopInformation_X86 should be updated after inlining
If function A has a loop and is inlined into function B,
HLoopInformation_X86 should be updated. Otherwise, Loop Peeling
phase into B tries to transform a the inlined loop, basing on
the information which belongs to the function A instead of B.
In other words, when A is inlined in B, nothing must refer to A
anymore.

Change-Id: I34a4fd1193d447abc4f38f4bb46bbdbeecd132fc
Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Upstream-Candidate: no, proprietary
Tracked-On: https://jira01.devtools.intel.com/browse/CAR-3373
Signed-off-by: Serdjuk, Nikolay Y <nikolay.y.serdjuk@intel.com>
Reviewed-on: https://android.intel.com/475754
Reviewed-by: Shamin, Anton <anton.shamin@intel.com>
Reviewed-by: Halimi, Jean-Philippe <jean-philippe.halimi@intel.com>
Reviewed-by: Mendell, Mark P <mark.p.mendell@intel.com>
Reviewed-by: Feng, Haitao <haitao.feng@intel.com>
Reviewed-by: Popov, Ivan G <ivan.g.popov@intel.com>
Reviewed-by: Lupusoru, Razvan A <razvan.a.lupusoru@intel.com>
Reviewed-by: Beyler, Jean Christophe <jean.christophe.beyler@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>