Basic Block profiling for JIT
Add support to the interpreters (goto & switch) to collect Basic Block
profiling while interpreting.  Save this information in the ProfileInfo
class in addition to the current invoke class types.

To figure out the BBs from the interpreter, when creating the profile,
figure out the target DexPC for each possible branch target.  This will
be used (later) to map the blocks with a given DexPc to the count for
that DexPc.

Ensure that IF fallthroughs trigger Branch events. This mirrors an
an AOSP change, but we can't use that due to other changes in that CL,
which are not in ART-Optimizer-N at this time.

Expected-Impact: none at this time.
Category: device enablement
Domain: AOSP.ART-Other
Origin: internal
Upstream-Candidate: no, proprietary
Tracked-On: https://jira01.devtools.intel.com/browse/CAR-3282
Change-Id: I3da1cb8058e20134ceefcd51245857c834a152c0
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>