ART-Extension: Add speculation framework + devirtualization pass
This patch creates the infrastructure for speculation framework and also
implements the first example user which is devirtualization. The current
implementation of devirtualization only handles precise types, but has
the infrastructure and the code to handle imprecise types - including
types that are obtained via profiling or CHA.

The speculation framework defines multiple recovery modes out of which
deoptimization, code versioning, and fault are supported. Code
versioning does not support counting because an infrastructure is needed
to actually support aggregating this data at runtime.

The speculation framework also defines multiple code versioning schemes.
Currently, only local versioning is supported which means that a guard
is added in front of the candidate in question. This is the least
aggressive technique since it does not bloat large code regions.

The speculation framework has the following initial limitations:
- Counting is not supported.
- Code versioning supported is only local.
- Grouped candidates do not use the same prediction.

Expected-Impact: Infrastructural item to support speculative
optimizations. It also adds infrastructure to support devirtualization
including possibly enhanced kind via CHA. In addition to that, it
supports precise devirtualization - this for example converts 3513
out of 9353 calls in IcyRocks from virtual/interface to direct calls.
In core-optimizing.oat, 40857/82232 calls are converted to direct. In
both cases, over 37% of resolved invokes are sharpened without guard.

Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Upstream-Candidate: no, proprietary
Tracked-On: https://jira01.devtools.intel.com/browse/CAR-2577
Change-Id: I99daccf74cfde6c061b1e789f9af148aff4794e8
Signed-off-by: Razvan A Lupusoru <razvan.a.lupusoru@intel.com>