ART-Extension: Do not clone HLoadClass in form bottom loops
Instruction simplifier for HCheckCast expects that second input
is HLoadClass while after form bottom loops it might be a Phi node
if HLoadClass is located in header but CheckCast is not.

To resolve it we do not clone HLoadClass node just because it
does not make sense. Clone will be moved to back branch basic block
which is dominated by pre-header where the original node is located.
So we just can use original node because there is no any sense to
load the same class twice.

Change-Id: I0e2e4deeec9bd263c4d3b9ecf6cbd505263c905a
Tracked-On: https://jira01.devtools.intel.com/browse/CAR-3506
Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>