Liveness: Handle non-first use equal to output
If the use of instruction is not its first input but
equal to first input and instruction requires the first
input is the same as output we can use the same position for
as we would be first input. It helps to avoid redundant movs on x86
for instructions like mul/add if both inputs are the same.

Change-Id: I83b78fdb86c887e4645ccd7f58a7e007436597b7
Tracked-On: https://jira01.devtools.intel.com/browse/CAR-3139
Category: device enablement
Domain: AOSP.ART-Other
Origin: internal
Upstream-Candidate: yes
Expected-Impact: Wide general improvement, impact is unknown at this moment
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>
Reviewed-on: https://android.intel.com/484190
Reviewed-by: Kazantsev, Maxim <maxim.kazantsev@intel.com>
Reviewed-by: Shamin, Anton <anton.shamin@intel.com>
Reviewed-by: Lupusoru, Razvan A <razvan.a.lupusoru@intel.com>
Reviewed-by: lab_aqa <lab_aqa@intel.com>
Reviewed-by: Mendell, Mark P <mark.p.mendell@intel.com>
Reviewed-by: Ivaschenko, Aleksei V <aleksei.v.ivaschenko@intel.com>
Reviewed-by: Halimi, Jean-Philippe <jean-philippe.halimi@intel.com>
Reviewed-by: Beyler, Jean Christophe <jean.christophe.beyler@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>