ART-Extension: Strengthen clinit motion check
In order to reduce impact of clinit check, Google's RA pushes the clinit
check to its users. This is the same concept as pushing null check to
the first user and thus do it implicitly. However, clinit check cannot
be pushed to user when there are other potential throwers or heap
mutators in between since the original problem would not occur in the
same program order as before.

So in order to prevent a problem and still allow for optimizations,
Google did the following things:
1) Added a conservative check that if the clinit check and its user
are in separate blocks, the check cannot be moved to user.
2) Added a sanity check for debug builds which checks that the
assumption holds.

However, we do more aggressive motion than Google. In the problematic
case, we actually fully unroll the loop and then merge the blocks. Since
redundant clinit checks get eliminated, if we have more than one user in
same block, we will fail Google's sanity check. Google did not consider
that the check would ever be in same block because they only apply LICM
which hoists to header - they never get rid of loops.

The sanity check is right. However, it needs to be tested in non-debug
builds to make sure we do not merge clinit check to its user in case
of aggressive code motion where we break the assumption.

Expected-Impact: Functional fix
Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Upstream-Candidate: yes
Tracked-On: https://jira01.devtools.intel.com/browse/CAR-3677
Change-Id: Ie02ad890d4bd8e9996ffd973d299bcbc3b9cc6ad
Signed-off-by: Razvan A Lupusoru <razvan.a.lupusoru@intel.com>
Reviewed-on: https://android.intel.com/487086
Reviewed-by: Zavjalov, Alexei <alexei.zavjalov@intel.com>
Reviewed-by: Kazantsev, Maxim <maxim.kazantsev@intel.com>
Reviewed-by: Katkov, Serguei I <serguei.i.katkov@intel.com>
Reviewed-by: Beyler, Jean Christophe <jean.christophe.beyler@intel.com>
Reviewed-by: Mendell, Mark P <mark.p.mendell@intel.com>
Reviewed-by: Halimi, Jean-Philippe <jean-philippe.halimi@intel.com>
Reviewed-by: lab_aqa <lab_aqa@intel.com>
Reviewed-by: Ivaschenko, Aleksei V <aleksei.v.ivaschenko@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>