ART: Update DCHECKs in SsaLivenessAnalysis::AddBackEdgeUses
Graph linearization in the presence of irreducible loops is not
guaranteed to generate a linear order where all blocks of a loop are
adjacent, first block is the header and last block is one of the back
edges.

These assumptions are made when inserting synthesized uses at the back
edges to aid the register allocator. Not meeting them will result in
the algorithm's early termination and back-edge uses not being added.

This patch updates the DCHECKs so the compiler does not fail in such
circumstances.

Bug: 27615840
Bug: 27624868

Tracked-On: https://jira01.devtools.intel.com/browse/CAR-3787
Expected-Impact: Functinal fix
Category: device enablement
Domain: AOSP.ART-Other
Origin: upstream https://android-review.googlesource.com/#/c/215892/
Upstream-Candidate: no, upstream
Change-Id: I63632e8819ea3644d5c6fdfea00b66128bf22c24
Reviewed-on: https://android.intel.com/496492
Reviewed-by: Mendell, Mark P <mark.p.mendell@intel.com>
Reviewed-by: Shamin, Anton <anton.shamin@intel.com>
Reviewed-by: Beyler, Jean Christophe <jean.christophe.beyler@intel.com>
Reviewed-by: Lupusoru, Razvan A <razvan.a.lupusoru@intel.com>
Reviewed-by: lab_aqa <lab_aqa@intel.com>
Reviewed-by: Fu, Chao-ying <chao-ying.fu@intel.com>
Reviewed-by: Katkov, Serguei I <serguei.i.katkov@intel.com>
Reviewed-by: Ivaschenko, Aleksei V <aleksei.v.ivaschenko@intel.com>
Reviewed-by: Popov, Ivan G <ivan.g.popov@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>
Reviewed-by: Halimi, Jean-Philippe <jean-philippe.halimi@intel.com>