ART: patchoat gets gzip support for compressed odex files
The patchoat is capable to inflate a compressed odex file
into a temporary file. Then the temporary file is patched
as if it is specified with --input-oat-fd or --input-oat-file.

The new options are the following:
  --input-oat-gz-fd
  --input-oat-gz-file
  --swap-fd
  --swap-file

Could not make in-memory decompression and patching because
the ElfFile class maps Elf file using its file descriptor. So
a temporary file is used as it is done by dex2oat.

Both gzip and zip are supported by zlib library and they have
the same compression algorithm. Gzip is chosen because zip needs
a zip entry name, which is unnecessary for single-file compression.

This patch was upstreamed to AOSP
https://android-review.googlesource.com/#/q/I54dbfcdc58bc0f57f7eccc3929c09295c597794b
but was rejected by Google. We decided to keep this set of ported
patches until a better compressed OAT support is implemented.

dalvik.system.DexFile.getDexOptNeeded() gets additional flag
DexFile.INFLATION_NEEDED returned by ART's oat_file_assistant
to denote that the dex file has its compressed odex file and
its APK file does not contain the original dex files. In this
case the APK cannot be run in JIT or interpreter mode without
the inflated odex.

Change-Id: Ib714438f97cf321d226a76c5dfb04735871159ac
Extension-Weekly: 1521a.15
Depends-Change-Id: I54dbfcdc58bc0f57f7eccc3929c09295c597794b
Depends-On: https://android.intel.com/503242
Depends-On: https://android.intel.com/503245
Depends-On: https://android.intel.com/503244
Depends-On: https://android.intel.com/507339
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-3209
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-32932
Category: aosp improvement
Domain: AOSP-Framework-CoreService
Origin: internal
Upstream-Candidate: no, not-aosp
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>
Reviewed-on: https://android.intel.com/503246
Reviewed-by: Halimi, Jean-Philippe <jean-philippe.halimi@intel.com>
Reviewed-by: Kanhere, Abhay <abhay.kanhere@intel.com>
Reviewed-by: lab_aqa <lab_aqa@intel.com>
Reviewed-by: Ivaschenko, Aleksei V <aleksei.v.ivaschenko@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>
Reviewed-on: https://android.intel.com:443/510581