Fix devirtualization issue with speculation guards
The current implementation of devirtualization doesn't always create the
right speculation guards. The implementation of IsPredictionSame()
currently considers two invoke predictions to be the same if the types
of the 'this' pointer are same. This is incorrect, as they may be
different instances of the same type.

Change-Id: I1d6b78191bff6ea6edb5ffe8f14fa06b76b79117
Extension-Weekly: 1521.11
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-3965
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-32427
Expected-Impact: none
Category: device enablement
Domain: AOSP.ART-Other
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>
Reviewed-on: https://android.intel.com/502697
Reviewed-by: Halimi, Jean-Philippe <jean-philippe.halimi@intel.com>
Reviewed-by: Zavjalov, Alexei <alexei.zavjalov@intel.com>
Reviewed-by: Lupusoru, Razvan A <razvan.a.lupusoru@intel.com>
Reviewed-by: lab_aqa <lab_aqa@intel.com>
Reviewed-by: Semukhina, Elena V <elena.v.semukhina@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>
Reviewed-on: https://android.intel.com:443/509291