Partial fragment deoptimization
We used to do either single frame deoptimization, or full fragment
deoptimization which deoptimizes all the frames in a fragment.
This change allows some methods to be not deoptimizeable, likely due
to some kind of optimization. So we need another deoptimization mode
that unwinds partial fragment. Deoptimizations are now generalized into
either full or partial fragment. A full fragment deoptimization will
deopt all frames in the fragment, and then returns from the invoke stub
to enter interpreter. A partial fragment deoptimization will deopt a
single frame, or all frames up to the method that's not deoptimizeable,
and then jumps to the interpreter bridge.

Currently code not deoptimizeable is the code in boot image since the
code may not be compiled with debuggable flag.

Bug: 28769520

Change-Id: I875c694791cc8ebd5121abcd92ce7b0db95aca38
Extension-Weekly: 1648.3
Tracked-On: https://jira01.devtools.intel.com/browse/CAR-4153
Category: device enablement
Domain: AOSP.ART-Runtime
Origin: upstream https://android-review.googlesource.com/233321
Upstream-Candidate: no, upstream
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>