ART-Extension: enable SILVER for methods with try/catch blocks
AUR traverses CFG and checks environment of each candidate instruction.
If environment refers to an insn which dominates the candidate then
environment is cleared. But later, prepare_for_register_allocation
may replace an instruction with its input which was cleared by AUR from
the environment. This corrupts stack map and leads to fault.

AUR should be disabled for 004-ReferenceMap because the test invokes a
native function which traverses stack and checks saved objects.
AUR should be disabled for 099-vmdebug because the test switches to
interpreter mode and tries to operate with wrong values that were not
saved in stack map because of AUR.

Change-Id: I0909684f7f78f852d4fc6b559d84db2a9e08f448
Extension-Weekly: 1629.7
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-3372
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-35621
Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Serdjuk, Nikolay Y <nikolay.y.serdjuk@intel.com>
Reviewed-on: https://android.intel.com/510745
Reviewed-by: Kazantsev, Maxim <maxim.kazantsev@intel.com>
Reviewed-by: Shamin, Anton <anton.shamin@intel.com>
Reviewed-by: Katkov, Serguei I <serguei.i.katkov@intel.com>
Reviewed-by: Mendell, Mark P <mark.p.mendell@intel.com>
Reviewed-by: Nesterov, Sergey <sergey.nesterov@intel.com>
Reviewed-by: lab_aqa <lab_aqa@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>
Reviewed-on: https://android.intel.com:443/523488