ART-Extension: Don't try to optimize phi inside catch blocks
Inputs of a CatchPhi don't dominate the catch block. If we remove
the DCHECK in phi_cleanup then we fail on the next DCHECK in
graph_checker because incoming phi was eliminated by phi_cleanup,
and, therefore, was not found in predecessor.

Change-Id: I2df4561f99df9c1791fac90a5e6b31f0b1c824e4
Extension-Weekly: 1631.9
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-4192
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-36278
Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Expected-Impact: performance changes are possible
Upstream-Candidate: no, proprietary
Signed-off-by: Serdjuk, Nikolay Y <nikolay.y.serdjuk@intel.com>
Reviewed-on: https://android.intel.com:443/527525