GenGC: Change the Parallel Copying implementation
We are fighting against fragmentation issue during
copying of "from" space to "to" and "old" spaces.
If we are using TLAB we are getting defragmentation issue
which can cause the case when there is no enough space in
"to" and "old" space to copy all live objects from "from" space.
It leads to fatal error and we want to avoid it.

To resolve the fragmentation issue we split the copying process
to two phases. The first one promotes the object to "old" space and
compute the required size in "to" space. On the second phase
we allocate the precise number of required bytes in "to" space
and copy objects there. This helps us to get a compaction
without fragmentation.

Change-Id: I88ffe310880c01dea53054abb67c127f4d1bda63
Extension-Weekly: 1637.9
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-4219
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-37771
Expected-Impact: unknown at this moment
Category: device enablement
Domain: AOSP.ART-GC
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Serguei Katkov <serguei.i.katkov@intel.com>
Reviewed-on: https://android.intel.com/535579
Reviewed-by: lab_aqa <lab_aqa@intel.com>
Reviewed-by: Ivaschenko, Aleksei V <aleksei.v.ivaschenko@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>
Reviewed-on: https://android.intel.com:443/538856