ART-Extension: Remove limitation from Sharpening
During enabling of Silver for methods with irreducible loops,
there was a crash on host build that was worked around by
disabling of pass sharpening2 for such methods.

The original reason of the crash was related with sharpening of
LoadString. The implementation of sharpening assumes that the
string being loaded is not present in dex cache. It is true
for the 1st run of sharpening, but is not true for 2nd.

Since Sharpening is the only pass that marks LoadStrings
as present in dex_cache, Sharpening2 can freely ignore such
LoadStrings, assuming that they have already been sharpened
during the 1st pass.

The fact that the problem happened in method with an irreducible
loop looks just a coincidence.

This patch restores scope of sharpening2 for methods with
irreducible loops (both JIT and AOT mode) and makes sharpening
to ignore the strings that have already be loaded.

Change-Id: I433b98d47654558024c76292f155e9e50ed790a8
Extension-Weekly: 1636.9
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-4257
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-37559
Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Expected-Impact: Sharpening2 applies for methods with irred. loops
Upstream-Candidate: no, proprietary
Signed-off-by: Max Kazantsev <maxim.kazantsev@intel.com>
Reviewed-on: https://android.intel.com/535910
Reviewed-by: Ivanov, Vladimir A <vladimir.a.ivanov@intel.com>
Reviewed-by: lab_aqa <lab_aqa@intel.com>
Reviewed-by: Semukhina, Elena V <elena.v.semukhina@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>
Reviewed-on: https://android.intel.com:443/537526