Check debuggable flag for quick deoptimized frames
When a quick (optimized) frame is deoptimized into an interpreter
(shadow) frame, it must contain stack map entries for all live virtual
registers. They must be valid for the dex code to be executed further
by the interpreter. This stack map correctness is guaranteed either by
the debuggable flag or HDeoptimize instruction. The latter is applicable
only for a single frame deoptimization.

This patch adds a check that for a multi-frame deoptimization each
frame corresponds to an oat method that was compiled as debuggable.

This patch invalidates many ART tests that deoptimize boot/core oat
frames because they are compiled as non-debuggable.

Change-Id: I901f991f14a05192249cd37b2d4e5c23dd6acc78
Extension-Weekly: 1649a.3
Orig-Tracked-On: https://jira01.devtools.intel.com/browse/CAR-4153
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-40707
Category: device enablement
Domain: AOSP.ART-Runtime
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>
Reviewed-on: https://android.intel.com/553284
Reviewed-by: Kazantsev, Maxim <maxim.kazantsev@intel.com>
Reviewed-by: Ivanov, Vladimir A <vladimir.a.ivanov@intel.com>
Reviewed-by: lab_aqa <lab_aqa@intel.com>
Reviewed-by: Shamin, Anton <anton.shamin@intel.com>
Reviewed-by: Semukhina, Elena V <elena.v.semukhina@intel.com>
Tested-by: Popov, Ivan G <ivan.g.popov@intel.com>