Currently, a HBoundsCheck is fed by an HArrayLength, causing a load of the array length, followed by a register compare.
Avoid the load when we can by comparing directly with the array length
in memory.  Implement this using an arch specific pass that replaces
the HBoundsCheck with a HX86BoundsCheckMemory that references the array
directly.  This allows us to generate code that references the length
in memory directly.

Only do this replacement when we are the only user of the ArrayLength
and it isn't visible to the environment.

Handle the special case where the array is 'null' and where an implicit
null check can't be eliminated.

This code moves the load of the length to the slow code for the failed
check, which is what we want.

Ported the following change Ids -
https://android.intel.com/#/c/597510
https://android.intel.com/#/c/611470
https://android.intel.com/#/c/610586

Change-Id: Ia5ebd06c2815cb33347cd3c0b3bdf622c5039a9b
Category: device enablement
Domain: AOSP.ART-Other
Origin: internal
Upstream-Candidate: No
Tracked-On:https://jira01.devtools.intel.com/browse/OAM-67390
Signed-off-by: ganesh shivaprakash <ganesh.shivaprakash@intel.com>
Reviewed-on: https://android.intel.com:443/639655