Fix on FaultHandler to distinguish native and oat methods better
Basically the x86 fault handler thinks that if the top of stack
contains a pointer that looks like an ArtMethod* (i.e. aligned and
contains a valid class pointer) then it is treated as such and
used to find its OatMethodHeader and translate native PC to dex_pc.
But there could be cases when the stack and native PC accidentally
satisfy the safe checks (alignmet and class verification) but still
correspond to a native method. In these cases behavior of ART is
undefined and further crash reports may contain misleading
messages.
The patch adds a safe check that the PC stays inside the address
range of the identified OatMethod method. In addition the PC check
in the stack walk routine is substituted with the safe
OatMethodHeader detection to not duplicate the code.

Ported original gerrit link:
https://android.intel.com/#/c/619170/

All tests are passed. No performance degradation in major KPI's.
Observed up to 2% performance improvement in jArtBand.

Change-Id: I58fba78ce5605da0961e36e0dbfbe2955922669a
Category: device enablement
Domain: AOSP.ART-Other
Origin: internal
Upstream-Candidate: no, proprietary
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-69437
Signed-off-by: crlalam <chinna.r.lalam@intel.com>
Reviewed-on: https://android.intel.com:443/642794